state: Move snapshot stale checks
Hunk navigation delegates batch file rendering, and selected-file snapshots
are written from the snapshot module.

Snapshot staleness checks still live in hunk navigation, so any
review-safety path that needs stale detection must depend on navigation
code.

This commit continues the file-review boundary split by moving snapshot
staleness checks into selected snapshot state and re-exporting the helper
from hunk navigation.

The next commit will point file-review state at the batch file renderer.
